Add step 0.1 to params and reload/reset buttons.

+++ description.py
@@ -5,7 +5,7 @@ import crseg.segmentation as cs
 import crdesc.description as cd
 import crdesc.config as cg
 
-def generateDescription(pigeon, uid, latitude : float, longitude : float, c0: int, c1: int, c2 : int):  
+def generateDescription(pigeon, uid, latitude : float, longitude : float, c0: float, c1: float, c2 : float):  
     G = u.Util.get_osm_data(latitude, longitude, 150, False, cg.way_tags_to_keep, cg.node_tags_to_keep, tempfile.NamedTemporaryFile(mode='w', delete=False, suffix=".xml", dir="cache/"+uid))
 
     # graph segmentation (from https://gitlab.limos.fr/jmafavre/crossroads-segmentation/-/blob/master/src/get-crossroad-description.py)
